+++ tasty-plutus/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0)
 
+## Unreleased
+
+### Changed
+
+* `tagged` is now `datum` to reflect its more general purpose.
+* 'Construction' functions in `Context` are now polymorphic over
+  `Purpose`s.
+
+### Removed
+
+* `ForCertifying` and `ForRewarding` constructors of `Purpose`. We have no plans
+  to support these.
+
 ## 2.0 -- 2021-09-14
 
 ### Changed